// resources/frontend/src/app/shared/models/expense.ts
import { User } from './user';
import { TypeExpense } from './type-expense';
import { Payment } from './payment';

export class Expense {
    id: number;
    nombre: string;
    fecha_fin: Date;
    flexible: boolean;
    user: User;
    id_tipo_gasto: TypeExpense;
    created_at: Date;
    updated_at: Date;
    pagos: Payment[];
}